requirements.txt
Django>=5.2,<6.0
djangorestframework
drf-spectacular
djangorestframework-simplejwt
django-filter
psycopg2-binary
pytest